refactor: 패키지 구조 변경
* common 패키지 추가 및 관련 파일 이동
