+++ BE/src/main/java/ifive/idrop/auth/filter/JwtAuthorizationFilter.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import ifive.idrop.entity.enums.Role;
-import ifive.idrop.exception.CommonException;
-import ifive.idrop.exception.ErrorCode;
-import ifive.idrop.exception.ErrorResponse;
+import ifive.idrop.common.exception.CommonException;
+import ifive.idrop.common.exception.ErrorCode;
+import ifive.idrop.common.dto.ErrorResponse;
 import ifive.idrop.auth.domain.AuthenticateUser;
 import ifive.idrop.auth.utils.JwtProvider;
 import io.jsonwebtoken.Claims;
