+++ BE/src/main/java/ifive/idrop/auth/filter/VerifyUserFilter.java
@@ -3,8 +3,8 @@ package ifive.idrop.auth.filter;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import ifive.idrop.dto.request.LoginRequest;
 import ifive.idrop.entity.enums.Role;
-import ifive.idrop.exception.CommonException;
-import ifive.idrop.exception.ErrorResponse;
+import ifive.idrop.common.exception.CommonException;
+import ifive.idrop.common.dto.ErrorResponse;
 import ifive.idrop.auth.domain.AuthenticateUser;
 import ifive.idrop.service.UserService;
 import jakarta.servlet.*;
