+++ BE/src/main/java/ifive/idrop/common/exception/CommonException.java
@@ -1,4 +1,4 @@
-package ifive.idrop.exception;
+package ifive.idrop.common.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
