+++ BE/src/main/java/ifive/idrop/fcm/FCMFilter.java
@@ -3,8 +3,8 @@ package ifive.idrop.fcm;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import ifive.idrop.auth.filter.JwtAuthorizationFilter;
 import ifive.idrop.dto.request.TokenRequest;
-import ifive.idrop.exception.CommonException;
-import ifive.idrop.exception.ErrorResponse;
+import ifive.idrop.common.exception.CommonException;
+import ifive.idrop.common.dto.ErrorResponse;
 import ifive.idrop.service.UserService;
 import jakarta.servlet.*;
 import jakarta.servlet.http.HttpServletRequest;
