+++ BE/src/main/java/ifive/idrop/parent/controller/ParentController.java
@@ -5,8 +5,8 @@ import ifive.idrop.dto.response.CurrentPickUpResponse;
 import ifive.idrop.dto.response.DriverListResponse;
 import ifive.idrop.dto.response.ParentSubscribeInfoResponse;
 import ifive.idrop.driver.domain.Driver;
-import ifive.idrop.exception.CommonException;
-import ifive.idrop.exception.ErrorCode;
+import ifive.idrop.common.exception.CommonException;
+import ifive.idrop.common.exception.ErrorCode;
 import ifive.idrop.driver.service.DriverService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
@@ -14,7 +14,7 @@ import java.util.List;
 import java.util.concurrent.ExecutionException;
 
 import ifive.idrop.auth.resolver.Login;
-import ifive.idrop.dto.response.BaseResponse;
+import ifive.idrop.common.dto.BaseResponse;
 import ifive.idrop.dto.request.SubscribeRequest;
 import ifive.idrop.parent.domain.Parent;
 import ifive.idrop.parent.service.ParentService;
