+++ BE/src/main/java/ifive/idrop/parent/service/ParentService.java
@@ -2,7 +2,7 @@ package ifive.idrop.parent.service;
 
 
 import ifive.idrop.driver.domain.Driver;
-import ifive.idrop.dto.response.BaseResponse;
+import ifive.idrop.common.dto.BaseResponse;
 import ifive.idrop.dto.request.SubscribeRequest;
 import ifive.idrop.dto.response.CurrentPickUpResponse;
 
@@ -10,8 +10,8 @@ import ifive.idrop.dto.response.PickUpHistoryResponse;
 import ifive.idrop.dto.response.ParentSubscribeInfoResponse;
 import ifive.idrop.entity.*;
 import ifive.idrop.entity.enums.PickUpStatus;
-import ifive.idrop.exception.CommonException;
-import ifive.idrop.exception.ErrorCode;
+import ifive.idrop.common.exception.CommonException;
+import ifive.idrop.common.exception.ErrorCode;
 import ifive.idrop.fcm.AlarmMessage;
 import ifive.idrop.fcm.NotificationUtill;
 import ifive.idrop.driver.repository.DriverRepository;
