+++ BE/src/main/java/ifive/idrop/pickup/controller/PickUpController.java
@@ -1,11 +1,11 @@
 package ifive.idrop.pickup.controller;
 
 import ifive.idrop.auth.resolver.Login;
-import ifive.idrop.dto.response.BaseResponse;
+import ifive.idrop.common.dto.BaseResponse;
 import ifive.idrop.driver.domain.Driver;
 import ifive.idrop.pickup.domain.PickUp;
-import ifive.idrop.exception.CommonException;
-import ifive.idrop.exception.ErrorCode;
+import ifive.idrop.common.exception.CommonException;
+import ifive.idrop.common.exception.ErrorCode;
 import ifive.idrop.pickup.service.PickUpService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
