+++ BE/src/main/java/ifive/idrop/pickup/repository/PickUpRepository.java
@@ -1,8 +1,8 @@
 package ifive.idrop.pickup.repository;
 
 import ifive.idrop.entity.enums.PickUpStatus;
-import ifive.idrop.exception.CommonException;
-import ifive.idrop.exception.ErrorCode;
+import ifive.idrop.common.exception.CommonException;
+import ifive.idrop.common.exception.ErrorCode;
 import ifive.idrop.pickup.domain.PickUp;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.TypedQuery;
