+++ BE/src/main/java/ifive/idrop/websocket/direction/NaverDirectionFinder.java
@@ -1,8 +1,7 @@
 package ifive.idrop.websocket.direction;
 
-import ifive.idrop.exception.CommonException;
-import ifive.idrop.exception.ErrorCode;
-import ifive.idrop.util.CustomObjectMapper;
+import ifive.idrop.common.exception.CommonException;
+import ifive.idrop.common.exception.ErrorCode;
 import ifive.idrop.websocket.direction.dto.Direction;
 import ifive.idrop.websocket.direction.dto.NaverDirectionResponse;
 import ifive.idrop.websocket.location.dto.Location;
