+++ BE/src/main/java/ifive/idrop/websocket/location/LocationWebSocketHandler.java
@@ -3,8 +3,8 @@ package ifive.idrop.websocket.location;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import ifive.idrop.driver.domain.Driver;
 import ifive.idrop.entity.*;
-import ifive.idrop.exception.CommonException;
-import ifive.idrop.exception.ErrorCode;
+import ifive.idrop.common.exception.CommonException;
+import ifive.idrop.common.exception.ErrorCode;
 import ifive.idrop.auth.domain.AuthenticateUser;
 import ifive.idrop.auth.filter.VerifyUserFilter;
 import ifive.idrop.auth.utils.JwtProvider;
